Add files via upload
A cut down version of the TMVAClassification.C file that focuses that trains and tests data based on the boosted decision trees and neural networks. Currently, the ability to choose which methods to use is limited to manually changing the booleans in the code, but this will be updated. Confirmed to work with ROOT versions 6.12.06 and 6.18.00. Confirmed to not work on ROOT 6.06, as this version lacks the dataloader.h file.