Add files via upload
Fixed some literary errors and booked an additional MLP method. Now the there are 2 methods that use the default arguments for the the number of hidden layers and number of nodes in each layer, but one uses the sigmoid activation function and the other uses hyperbolic tangent. The same applies for the other two MLP methods, but their network architecture is specified with both having 2 hidden layers, the first layer with 17 nodes, and the second layer with 16 nodes.